+++ CHANGELOG.md
@@ -1,10 +1,16 @@
+## 1.2.0
+
+- Updated web implementation.
+- Update SDK constraints to `>=2.14.0 <3.0.0`.
+- Update `test` package constraints to `^1.18.0`.
+
 ## 1.1.1
 
 - Update meta.
 
 ## 1.1.0
 
-- Add a `value` parameter to `ComputeOperation<R>.cancal()`.
+- Add a `data` parameter to `ComputeOperation<R>.cancal()`.
 
 ## 1.0.2